backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
# package-modeをfalseに設定
package-mode = false
[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
firebase-admin = "^6.5.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.31"
psycopg2 = "^2.9.9"
alembic = "^1.13.2"
openai = "^1.35.10"
stripe = "^10.2.0"
pandas = "^2.2.2"
asyncpg = "^0.29.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
sqlalchemy = "^2.0.31"
aiosqlite = "^0.20.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
